Fixed Refresh Token Save
Refresh Tokens, can be updated if you do not want the Refresh Token to
rotate when the access token expires. Made changes to allow for this,
and did some clean up with some functions and cleaned up the Grant
flows. Should be good to commit as v1. Tested working in Unbox.